cli/requirements.txt
# DRRMS CLI Dependencies
click>=8.0.0
mysql-connector-python>=8.0.0
tabulate>=0.9.0
python-dotenv>=1.0.0
